Do not attempt to read from non-existing input stream
This happens at the end of the input when no more data is available
but the data has not actually been read from an input stream (and thus
the input stream is empty). Note that in this case the buffer is not
completely empty (and if there was a file and it was EOF, no error
would have been thrown, too).